enhancement: Use generators instead of lists and remove extra calls

+++ hotpdf/helpers/nanoid.py
@@ -17,7 +17,7 @@ def method(algorithm: Callable[[int], bytearray], alphabet: str, size: int) -> s
     mask = 1
     if alphabet_len > 1:
         mask = (2 << int(log(alphabet_len - 1) / log(2))) - 1
-    step = int(ceil(1.6 * mask * size / alphabet_len))
+    step = ceil(1.6 * mask * size / alphabet_len)
 
     id = ""
     while True:
